Add SPARQL* support for direct (non-reified) modeling of embedded triple patterns.
Adds new Algebra.matchStatement to represent the matching of a triple pattern
and the binding of its identifier to an (internal) variable.

Changed SPARQLParser from a struct to a class, and added a SPARQLStarParser
subclass that sets its `parseSPARQLStarAsRDFReification` state to produce
Algebra structures that model an embedded triple as the join of a matchStatement
and a triple pattern. The SPARQLParser base class is still able to parse SPARQL*
syntax, and will continue to produce algebra results using RDF reification
patterns.